Cursor for KeyFrameWidget can be changed by dragging left button across window - keyframes added are placed at cursor position

+++ src/widgets/keyframewidget.h
@@ -52,8 +52,8 @@ private:
     float minRange;
     float maxRange;
 
-    //indicates current time
-    int cursorPos;
+    //indicates time at cursor
+    float cursorPos;
 
     bool dragging;
     int scaleRatio;
